Split out test
Many tests and variant and tests running serially inside a class
makes this the test class that finishes last when running. Split out
du to that and since the comment indicates it belongs in its own
class anyway
